Updated BattleTrigger and BattleManager
Now saves the instance of the latest BattleTrigger that starts a battle as to not send the player reference around so much in the code.

BattleManager now calls on the BattleEnded() method in BattleTrigger to notify the end of the battle. So all logic for the player and enemy GameObjects (in the scene) during battles are now handled by BattleTrigger.

The enemy and its trigger objects now disappear from the game when they are defeated.

The battle logic should be mostly done, though nothing is shown to the player yet.